+++ tests/Server/Handler/Request/CompleteRequestHandlerTest.php
@@ -146,7 +146,7 @@ final class CompleteRequestHandlerTest extends AbstractMcpTestCase
      *     ref: PromptReference|ResourceTemplateReference,
      *     argumentName: string,
      *     argumentValue: string,
-     *     contextArguments: null|array<string, string>,
+     *     contextArguments: null|array<array-key, string>,
      *     context: ServerContext,
      * }
      */